+++ styles/components/multimedia-slider/FullScreenCarousel.module.sass
