+++ styles/components/multimedia-slider/GoFullScreenButton.module.sass
