+++ styles/components/multimedia-slider/MultimediaSlider.module.sass
